# swagger_old_old.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RTQL_Backend_Service",
    "version": "1.0.0",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:3677/v1"
    },
    {
      "url": "http://64.181.233.131:3677/v1"
    }
  ],
  "paths": {
    "/signup": {
      "post": {
        "summary": "signup for real",
        "tags": [
          "signup"
        ],
        "responses": {
          "200": {
            "description": "signup for real",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "default": "<integer>"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fname": {
                    "type": "string"
                  },
                  "lname": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "pass": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "login for real",
        "tags": [
          "login"
        ],
        "responses": {
          "200": {
            "description": "login for real",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "default": "<integer>"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "pass": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/question/{qid}/response": {
      "get": {
        "summary": "get answers",
        "tags": [
          "question",
          "response"
        ],
        "responses": {
          "200": {
            "description": "get answers",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Date": {
                "schema": {
                  "type": "string"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "retrieves the answers (responses) for a particular question."
      },
      "parameters": [
        {
          "name": "qid",
          "in": "path",
          "required": true,
          "deprecated": false,
          "schema": {}
        }
      ]
    },
    "/question/{qid}/response/{id}": {
      "put": {
        "summary": "update question answer",
        "tags": [
          "question",
          "response"
        ],
        "responses": {
          "200": {
            "description": "update question answer success",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Date": {
                "schema": {
                  "type": "string"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "response": {
                    "type": "string"
                  },
                  "rid": {
                    "type": "integer"
                  },
                  "correct": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "qid",
          "in": "path",
          "required": true,
          "deprecated": false,
          "schema": {}
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "deprecated": false,
          "schema": {}
        }
      ],
      "delete": {
        "summary": "delete question answer",
        "tags": [
          "question",
          "response"
        ],
        "responses": {
          "200": {
            "description": "delete question answer success",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Date": {
                "schema": {
                  "type": "string"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/question": {
      "post": {
        "summary": "add question",
        "tags": [
          "question"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "deprecated": false,
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "save question",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "default": "<integer>"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              }
            }
          }
        },
        "description": "saves a question for the user who is currently authorized.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string"
                  },
                  "options": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "correct": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "get question",
        "tags": [
          "question"
        ],
        "parameters": [
          {
            "name": "qtext",
            "in": "query",
            "required": false,
            "deprecated": false,
            "schema": {}
          },
          {
            "name": "verbose",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "deprecated": false,
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "get question",
            "headers": {
              "X-Powered-By": {
                "deprecated": false,
                "schema": {}
              },
              "Access-Control-Allow-Origin": {
                "deprecated": false,
                "schema": {}
              },
              "Content-Length": {
                "deprecated": false,
                "schema": {}
              },
              "ETag": {
                "deprecated": false,
                "schema": {}
              },
              "Date": {
                "deprecated": false,
                "schema": {}
              },
              "Connection": {
                "deprecated": false,
                "schema": {}
              },
              "Keep-Alive": {
                "deprecated": false,
                "schema": {}
              }
            }
          }
        },
        "description": "retrieves the questions for the user who is currently authorized."
      }
    },
    "/question/{id}": {
      "put": {
        "summary": "update question",
        "tags": [
          "question"
        ],
        "responses": {
          "200": {
            "description": "update question",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Date": {
                "schema": {
                  "type": "string"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "use this question to update a question as the user who is currently authenticated.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "qid": {
                    "type": "integer"
                  },
                  "question": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "deprecated": false,
          "schema": {}
        }
      ],
      "delete": {
        "summary": "delete question",
        "tags": [
          "question"
        ],
        "responses": {
          "200": {
            "description": "delete question",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Date": {
                "schema": {
                  "type": "string"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin": {
      "get": {
        "summary": "admin stats",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "admin",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "default": "<integer>"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              }
            }
          },
          "403": {
            "description": "not an admin",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "default": "<integer>"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "default": "<string>"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
        "security": [
            {
                "bearerAuth": []
            }
        ]
    },
    "/model": {
      "post": {
        "summary": "new question",
        "tags": [
          "model"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "deprecated": false,
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "new question",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Date": {
                "schema": {
                  "type": "string"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Use this endpoint to get questions from the model.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "input": {
                    "type": "string"
                  },
                  "questions": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}